pyproject.toml
[project]
name = "credtech"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "db-sqlite3>=0.0.1",
    "dotenv>=0.9.9",
    "feedparser>=6.0.11",
    "finnhub-python>=2.4.24",
    "matplotlib>=3.10.5",
    "mcp-yfinance-server>=0.1.0",
    "nltk>=3.9.1",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "regex>=2025.7.34",
    "requests>=2.32.4",
    "seaborn>=0.13.2",
    "sec-api>=1.0.32",
    "websocket>=0.2.1",
]
